Merged PR 395: 4726 - part 2 | CCU Log file upload
This change keeps the the ".zip" on the uploaded file name to make it easier to process by the end user, as requested in ticket by QE.

I left the .zip off initially as I read the requirements for the fileID in the upload to literally.   The API works fine with .zip.

Related work items: #4726