Merged PR 389: 2679:  Align CCU "kind" tags to correct data types and move to Silo
The main changes in this PR are

1) Pointing to the new **Silo** service endpoint in place of remote Haystack,
2) In Haystack tags, changing all Kind: "string" to the proper **Kind: "Str"**, to be consistent with the spec, and
3) Adding a new dev feature to **pre-populate** the site fields in a **new registration**.

### 1. Silo Endpoint

After changing the endpoint to the new service, verification is required and has so far consisted me
* perusing the logs for exceptions and response code 500s,
* examining Silo via the portals,
* examining Silo directly with Postman,
* using portfolio analytics manager to look at plots, since the portals are still pointed at Haystack and are no longer getting his data.

A couple of issues were found and fixed:
1. I found and fixed one issue here ->  the mime type of the `Http POST` to remote Haystack in `HClient` was setting an unsupported mime type, being rejected by Silo.  It's now `text/zinc`
2. I removed the line in `ScheduleProcessJob` that writes historical data to the occupancy status message point.  The status message is a string and this point is given a string default value, whereas the occupancy his values are numbers.  This is not be supported by Silo, although removing this may create an issue since the actual his point may not be created properly.  See https://dev.azure.com/75fdevelopment/75F%20Engineering/_workitems/edit/5000

Otherwise, this change has passed the smoke test, although additional issues could certainly arise.

### 2. Str conversion

To be consistent with the latest Haystack API, as supported by Silo, "kind" tags with a value of "string" or "String" needed to instead have a value of "Str".   This PR does that and also adds a one-time migration block to transform existing tags in the Haystack database.

This one-time migration occurs at startup in the init of the `CcuApi` class.  The migration ensures the updated tags are saved to disc and synced with the server.

### 3. Pro-populate registration in Dev

Purely as a convenience, I created a feature which auto populates a new site during site creation.  It currently works only in the local and dev environments.  If you including the following property in your local.properties file, the site information page will pre-populate every field except for the site name.

`debug.user.email="[your email]"`

### Other Notes

* Earlier we found one issue with the portal, which I documented with a new ticket: https://dev.azure.com/75fdevelopment/75F%20Engineering/_workitems/edit/4915 (Statuses display only as "Loading...").  However, the task of switching the portals to point to Silo has not yet been completed, and we'll have to see what additional issues come up there which may affect CCU.
* Also included in this PR are minor additions and tag improvements to some of our logging statements