Merged PR 457: 5141  Dual duct points are not appearing in Portal Zone UI
Dual Duct appears broken when the profile is used after the Silo migration.

Turns out the profile had many points erroneously marked as `kind:"string"`.
Haystack didn't mind that but they Silo enforces the spec.

This PR updates those to not have a String and fixes DualDuct.  About 15 points.

Another point -- DualDuct's `scheduleStatus`, was a point used both as a string with a default value, and as a his with numeric values..  I removed the his in that case.

I also updated `HaystackExplorer` in two ways
1) do not require a password to change a value on Local or Dev,
2) do not call `updateData()` in the expand click listener, since that was very occasionally causing the backing data to change -- with the sections being reordered -- and causing a crash with an `IndexOutOfBounds` exception.  @<samjith>  we can change that back if you need it.

Related work items: #2679