Merged PR 481: Unit Tests Re-intro
Ticket:  https://dev.azure.com/75fdevelopment/75F%20Engineering/_workitems/edit/5218

The main point here is to start up unit tests again.

NOTE: that this PR also includes a simple enhancement to data migration -- an oversight was causing the migration to occur every time the app started up, instead of only the first time.  Resulting in unnecessary processing.

This work:
* Adds some unit test libraries to the haystack module.  (same can be done elsewhere)
* Includes a unit test of the existing data, haystack to silo format migration function.  This can serve as a sample Kotlin unit test, although it's pretty simple.
* Cleans up past unit tests, many very old and either not really unit tests or failing.

You can now run unit tests as part of gradle:

`./gradlew clean testDev`

This should be done prior to pushing a branch for a PR to ensure all tests are working, and with all tests passing.  In the future, this will be run as part of CI/CD to ensure passing tests.

It would be an excellent practice to require some unit test as part of each new PR.