Merged PR 599: Revert 'Alerts V2 Endpoints cutover'
Reverts !556

This is necessary as this change should not go into a release without also the Alerts-v2-pubnub change.   While that branch could be merged today, it is quite unlikely it could be verified quickly enough to make this Sprint's release.

Hopefully a release will be cut and we can merge this back in as soon as tomorrow.

Related work items: #5264, #5280