Merged PR 616: CCU Crash reporting
This work adds crash reporting back to the CCU app.  Since Crashlytics was still not working for us after much investigation, we have gone with another service called Raygun.

On top of very basic configuration, added a version name, and a user name, in the form of siteName: + ccuName to each report.