Merged PR 975: Delete floor remotely only if not used by other CCU
Whenever a `floor` is deleted by the user, we now check our remote site on the server.  Specifically, we check for `equips` that match our `floorId` but have a different CCU (a different `ahuRef` or `gateway` ref).  If so, we only delete the `floor` locally.
If not, we're clear, and delete the `floor` on the server as well.

I added a wait spinner for when all the processing and server call are occurring.   I use RxJava to implement the aync parts.  If anything untoward happens, e.g. no network, and we come back from the RxJava in the `onError` handler, then the `floor` is not deleted and I put up a Toast to that effect.