Merged PR 2763: Bug 9727: Hyperstat User Intent Override Issue
When we move to auto force occupy then we need to update next schedule desired temp when zone following zone schedule.
When we are in auto force occupy if desired temp is changed from the portal and if we are extending the time then we need to consider top level desired temp value.

Related work items: #9727