Merged PR 2826: 9728 : CCU - Zone schedule desired temp update Changes
If multiple schedule is added in zone level , any  schedules other than the first schedule is not updated  with desired temp change made from portal and is overridden by the default schedule desired temp .
This was caused because the desired temperature was updated without checking the specific schedule and updated to the first found schedule of that day.
Now we are comparing the current time with the schedule time and only then update the schedule.

Related work items: #9728