Merged PR 3021: 10108 - Prevent NPE caused by bad custom alert defs
If a custom alert definition contains a message "token" (i.e. #pointVal, #condVal) that doesn't have a matching value, the template parsing fails and the alert occurrence is never created. This simply prevents the NPE and fills in a blank value so the occurrence can still be created and sync'd

Related work items: #10108