Merged PR 3489: Reboot msg has been updated
When user performs "reboot" operation from portal a control message is sent to node devices. The field "reset" in the control message is set to 1 to indicate node devices to reboot. However, when this control message is sent, the CCU sets all the other fields in the control message to zeroes. Now we are reading the control message with all other fields set.

Related work items: #10706