Merged PR 6896: 18179 17552 Run Fan stages only when we have conditioning
Fan stages should run only when conditioning is happening. When supply and demand are the same then the fan should operate.
When supply and demands are different and if the zone asks for heat then the aux stages will turn on, Also Fan stages should run according to the aux stages.

Related work items: #17552, #18179