Merged PR 6721: US - 16069
Crash 1: As of now, we have included a try-catch block because the UsbSerialDevice (UsbSerial library class)
is invoking the connection.requestWait() method within a while loop. If a NullPointerException is encountered during a particular iteration,
 the program will attempt to execute the method again in the next iteration.

We looked is there any new updates UsbSerial library, as of now library updated 4 years ago
https://github.com/felHR85/UsbSerial/tree/319cb9de686325dcc2814d7cbb6ed7dd36011183/usbserial/src/main/java/com/felhr

To handle this safely, we have implemented a try-catch block in this section of the code.

Crash 2: The exception occurs only when the app is not installed in the system/priv-app folder,
and does not occur under normal circumstances.  - So no fix provided for this As of now.

Related work items: #16069