Merged PR 8262: auto-cx aborted state handled now-mstr
Handled below scenario in this PR

Scenario 1: Scheduled for Auto Commissioning test to run for 1 hour starting at 14:00.
When the test is aborted at 14:05, the status shows that the test was aborted for a minute and then the next minute shows that it is not started.
14:00 - Test started
14:05 - Aborted
14:06 - Not started

Scenario 2: Scheduled for Auto Commissioning test to run for 1 hour starting at 14:00.
The test runs completely for the scheduled hour. Status should show that the test is completed at the scheduled stop time and the next min should show that it is not started.
14:00 - Started
15:00 - Completed
15:01 - Not started

Related work items: #20516, #21232