Merged PR 8105: DaikinIE reconf - Enabled point is updated for every point update with min max values.
Currently we are setting the "enabled" point for all the config point update. That is if there is a change in coolingMin, we are updating the "enabled" point with the value that is sent. This is writing "enabled" point with values set to min/max.
Removed the write as it gets updated with updatepoints

Related work items: #20745