Merged PR 9064: removed his value write operation.
removed writeHisVal operation.

In the TI profile, we have 3 sensors(`sensor bus temperature`, `thermistor input1`, `thermistor input2`) to write a value to the `current and temp` point so, whenever we have enabled the TH2 and then enable any other sensor(`sensor bus temperature` or `thermistor input1`) we were writing the value 0 to the `current and temp` point irrespective of whether the TH2 is enabled or not. This was causing the issue, so there was a miss in the code, corrected the code to only write a value to the point whenever the sensor is enabled and the value is changed.

Related work items: #20560