In general, HyperStat relays are controlled as follows for staged outputs:
    1. If loopOutput > threshold + relayActivationHysteresis/2, enable Relay
    2. If loopOutput < threshold - relayActivationHysteresis/2, disable Relay
    3. If loopOutput falls between the previous two cases, keep Relay at its previous value

This is the correct logic for the algo, but status message was not following it in Scenario (3). In this case, status message always excluded the stage, even if it was still on due to being enabled previously.

In the third case, relay status was always excluded from status message previously, even if the relay was still on due to being enabled before. Status message has now been fixed in HyperStat classes to accurately reflect relay state.