writeControlMessage() method was writing the hash value of the control message to the message cache entry for settings message (not control message).
This was causing the settings message to re-send every minute because the settings message hash was always different from the one most recently written to the cache. This bug was present for HyperStat and HyperStat Split.

Controls message is now written to its correct location in the message cache, and settings message now only sends upon a change in contents.