Hacks to improve responsiveness/speed of config UI.
There is a render() method that updates the UI upon a user action. On HyperStat profiles, this method reloads the entire UI upon every action. With Split, there are enough points that this leads to performance issues if the user clicks too fast.

This UI code in HyperStatSplitFragment has been updated to only reload UI items relevant to the change made. Performance is improved significantly with this approach.

The approach used is not clean, but this section of UI will be upgraded to Jetpack Compose before being built on further.