Merged PR 12856: Bug/Boopathy/Bug_27737_CCU_Airflow_volume_value_is_not_writing_in_truecfm
The airflow cfm value in the DAB profile is showing zero even after the calculated value is seen in logs. The value is written in different set of tags where the point is not available. Changed the sensor tag to trueCfm tag and made it to write in the concerned point.

Related work items: #27737