Merged PR 13669: 29247: buildingLimitMin, buildingLimitMax and buildingToZoneDifferential points are not updated to newer format as part of cutover migration - Master
syncing building limit min/max and building to zone differential points from cloud upon version upgrade even if the points does not have domainName

After pulling from latest master, some issues were observed due to which migration was not working. I have fixed that issue in this PR as well. In MigrationHandler, we call a method to udpate the ahuRef of TI Profile. Regardless, TI profile is paired, we were trying to update the ahuRef which was causing NullPointerException.
I have added a check to see if the equip existis or not and based on that continue with the below code.

Also some logs were added in format Log.i() instead of CcuLog.i() without even importing the required modules which resulted in compile time error.

Related work items: #29247