Merged PR 14410: 30071: otaStatus update should consider only the last bit sent in the nodeStatus message from CM to CCU - Master
consider only last bit for messageType

NOTE: @<Manjunath Kundaragi>  In my first commit, I added the change in only 1 out of 2 methods which parses the last 3 bits. Due to this qa testing was not possbile. In my recent commit, I have made the changes in the second method as well.

Related work items: #30071