Merged PR 15756: 31391 - Home app OTA plus RC script installation
Cherry picked merge from master_hotfix_2.14

This PR includes both the changes for Home App OTA (including tablet reboot) as well as the installation of RC initialization scripts. Look in assets/init/README.md for more.

Note: There is still an opportunity to implement a failsafe in the form of a low level background process to monitor the CCU app and bring it forward if it is in the background or not running. The ability was demonstrated to the ADR group, but it was decided at that time to focus on HomeApp OTA cleanup instead. This background process DOES NOT suffer from the same limitations that the HomeApp is constrained to and would be a much more reliable way to keep the CCU app in the foreground.

Visit ADR-0105 to learn more....

https://dev.azure.com/75fdevelopment/75F Engineering/_wiki/wikis/75F-Engineering.wiki/3537/ADR-0105-CCU-system-initialization-and-monitoring-at-tablet-boot-time

'nuff said...

Cherry picked from !15700

Cherry-picked from commit `9be3729b`.

Related work items: #31391