Merged PR 15872: 32051 - When users pair a Bypass damper, close and relaunch the app, CCU is crashing - master
State changes in Jetpack Compose need to happen on the main thread, especially when dealing with UI updates. If a state is modified in the background (e.g., from a coroutine running in Dispatchers.IO), Compose may not apply that state change correctly to the UI, causing this error.

Root cause:

If you update the state in an asynchronous background thread (like a coroutine) and read it before Compose can apply the change, it leads to inconsistency between the state snapshot and the actual state

Fix:

i have updated to the highest priority thread to do the viewmodel init   first before call the rootview

Related work items: #32051