Merged PR 16233: 32255 - Adjusted HomeApp OTA process to account for older filenames - Master
Adjusted HomeApp OTA process to account for older filenames.

It turns out that there are a couple of apk patterns that need to be looked for:

- 75fHomeRenatus_v3.1.apk (we were dealing with this one)
- HomeRenatus.apk (this one was a surprise)

This PR deals adjusts the OTA process, including the 75f-init.sh script to deal with HomeRenatus.apk if it is present as well.

Cherry picked from !16214

Cherry-picked from commit `cad51fb8`.

Related work items: #32255