Merged PR 18068: 34203: Point count before un-register & after re-register of the CCU are not matching when OTN is paired - Master
#34203

Corrected the check to find building points.
Point class, does not store ccuRef in tags list. It is a direct attribute.

The missing points are OTN specific **tuner** points with **default** tag. When updateEntity message is received post cutover migration, CCU is assuming it as a buidling point and updating the buildingEquip's id as their equipRef.
Since, the tags list does not contain ccuRef and the point contains **"default"** and **"tuner"** tag, it is assumed as Building level point.

**NOTE**:
I tried different migration scenarios, to see if any other points other than the OTN related are being affected.
Following point had tuner and default despite not belonging to building tuners:
- **OTN**: (since DM migrated)
   - Proportional KFactor
   - Integral KFactor
   - Zone Priority Spread
   - Zone Priority Multiplier
   - Temperature Integral Time
   - Temperature Proportional Range
- **Hyperstat**:
   - standaloneMinFanRuntimePostConditioning
- **Vav Fully Modulating AHU**:
   - Reheat Relay Activation Hysteresis

I tried migration from 2.7.4 to 2.13.9 to 2.20.3, 2.7.4 to 2.20.3

Only the OTN tuner points are getting their equipRef udpated to buildingEquip's id and somtimes a few of them are getting deleted due to syncBuildingTuner operation.

I think, we may not need to give a migration fix since DM OTN was released in 2.20.0 and no production release have been approved as of now.

Related work items: #34203