Merged PR 18990: include cm-currentTemp in system avg temp calculation when OTN is paired - master
Include cm-currentTemp in system avg temp calculation when OTN is paired

Issue: When OTN is paired (without CCU-TI) along with other profiles, the cm-current temp was not included in the calculations for system avg temp, weightedAverageCoolingOnlyLoadSum, weightedAverageHeatingOnlyLoadSum, and weightedAverageHeatingConditioningLoadSum.

Fix: When OTN is paired (without CCU-TI) along with other profiles, include the cm-current temp value in the calculations for system avg temp, weightedAverageCoolingOnlyLoadSum, weightedAverageHeatingOnlyLoadSum, and weightedAverageHeatingConditioningLoadSum.

Related work items: #35318