Merged PR 18981: syncing unSynced log files - master
Issue:
When a user requests log files (cculogs, sequencer logs, or alertlogs) by sending a remote command from the portal, the CCU is expected to upload these logs to the cloud. But the main reason for the failure is a slow/no internet connection, which causes the logs unsynced and unavailable.

Fix:
Now, the CCU will try to upload the log files that were already created once the internet connection becomes stable.

Related work items: #35079