Merged PR 20070: 36210: CCU: Backfill Time Reset to NONE After Migration to 3.2.1 - master
changed the migration flow and added code fix to update the default value

RC :
During the migration from version 2.7.4 to 3.2.1, the backfill point was originally present in the system equipment in 2.7.4.

However, during the cutover migration in 3.2.1, if the migration process occurred for the system equipment, the backfill point was not found in the model and was removed since it had already been migrated to the CCU base configuration.

When the CCU base configuration is applied, we attempt to read the point’s default value, but since the point itself is missing, the default value returns 0. As a result, we write this value to the CCU-based backfill point

Related work items: #36210