Merged PR 20726: Bug 37381: CCU: Outside Temp Lockout Cooling Toggle Gets Disabled After Connect Module Add/Remove or Reconfiguration Without OAO-master
RC : We are not checking whether the Connect module OAO is paired or not. That’s why, if the user pairs only the V2 CM, the system-level OAO is considered paired. When the user performs a reconfiguration, it is disabling the Cooling Lockout toggle.

Correction: added the check if the connect module is paired and oao is not paired in the system level; then, based on the use cases, enabling and disabling the cooling lockout

Related work items: #37381