Merged PR 21202: 37781: CCU: Conditioning mode is not displaying the expected values in drop down section-master
updated the hiswrite val when we update the config

RC : While running the profile algorithm based on the hisValue, we update the default value. However, during reconfiguration cases—when we change the conditioning mode—we are not writing the updated hisValue. As a result, when the profile algorithm runs, it uses the old hisValue, which leads to an incorrect update of the default value.

correction : writing the his value in reconfiguration cases also

Related work items: #37781