Merged PR 22253: side app not installed fix - master
side app not installed fix - master

Scenario:
Whenever we have home, remote, and the BACApp app to update... Since updating the remote app and BACApp app, we were using the code that was running in the background. As soon as Home is installed, it reboots the system, and the BACApp/remote app was not installed.

Related work items: #38495