Merged PR 22525: time zone change update schedule/event entity
When user changes the **timeZone** from the ccu settings page and save it. CCU is updating all the points, during this update, some of the properties in point schedule/event entity was updated with "" hence this was creating issue while applying the custom value to the point and hence it was showing **NA** status.

Related work items: #39315