Merged PR 23486: null check to prevent subsequent migration - master
Issue 1: The HyperStat migration failed(issue2). Because of this, the model migration code did not run, and the CCU version was not updated.

Issue 2: From the logs, I can see that one HyperStat migration failed. It was trying to fetch the relay1 entity using the domain name, but it failed since the relay1 point itself has no domain name. In version 3.0.3 (specific to this CCU), I can still see some device points without domain names, which means they are not migrated earlier.

Fix for Issue 2: When no device entity is found, we currently check port.isEmpty. We will also add a null check here to prevent a crash and avoid issues in the subsequent migration.

Issue 3: This happened because the system profile was earlier set as VAV STAGED RTU, but it should now be VAV STAGED RTU WITH VFD. Since the HyperStat migration failed, the migration from Staged to VFD also failed.

Related work items: #40410