Merged PR 22816: 39627,40235: For_Modbus_points_which_are_schedule_is_not_assigned_but_schedulable_point_is_there_user_intent_values_are_overridden_by_level_17 - Master
#39627
#40235

Changed the code to clear priority array when schedulable tag is removed during local/remote reconfiguration or when a point following schedule has its scheduleRef removed or when point gets its both scheduleRef and eventRef removed.

Whenever data is received for any modbus register, the corresponding logical point's highest priority data will be sent back to end device if the received data is not matching with highest priority data.
Point based scheduling operation will not trigger the operation to send the point's value to end device for modbus unless an active event is getting expired the next minute.

For Bacnet Client profile, schedular can immediately send the data.

If the external profile equip's logical point has the schedulable, scheduleRef or eventRef tag removed (while scheduleRef is not present), then level 8 value which is controlled by the custom control by now needs to be cleared out and the highest prioirty data needs to be updated to the end device. This needs to be done since the point based scheduling cannot happen with the point and point's default value should be shared with the end device unless any other override is happening.

**NOTE:**
1. I have added some comments upon all possible code changes made by me explaining the reasons.
2. Fix for following observations found while addressing the issue have been added:
   - While sending the data for Register points to end device with no wordOrder tag and parameterDefinitionType set to float was failing as the code tries to access word order and if it matches to bigEndian or littleEndian, thus causing NullPointerException which is soft handled.
   - Physical modbus points with writable tag were not getting updated even though the logical point's value was updated. This was causing his value and highest priority value to  mismatch.
3. While I have address both issues. If duration is provided for any array level for a logical point, CCU intrinsically clears out the array level whenever the point's writable data is read. But this will not automatically update the physical points. I did not make any changes for this one considering that it is a common code.

Related work items: #39627, #40235