Merged PR 23981:  40751: [CCU] Stuck in Continuous Loading After Clicking “Finish” During Preconfigure Site Registration.(iProda) - master
add the code to create ccu device

Found the RC :

When we start the Sync, we first check whether the site is already synced or not. If it’s not synced, we attempt to fetch it and then sync it to the cloud.
During the Pre-config, we try to create a point and call the method to sync it to the cloud. At that moment, before syncing any point, we again check if the site is synced. If not, we attempt to sync it, and if the site doesn’t exist, we try to create a site.

Once we navigate to the Congrats fragment, if the Finish button is clicked, we check again whether the site is synced.

If not, we attempt to create a site, and then we create a CCU.

If the site is already synced, we skip CCU creation — and the loader is not being closed.

Use case causing the issue:
The site was created and synced during the Sync process. When the user clicks Finish, the site is already synced, so the CCU is not created, and the loader remains open

Related work items: #40751