Merged PR 24336: 41259: CCU discrepancy unable to change the unoccupied zone setback value to non default values - master
unoccupied set back point level value

When the _Follow Building_ toggle is disabled, we now fetch the **Unoccupied Setback** value from the **zone-level point**.

However, when clicking **Save**, the Unoccupied Setback value is updated **only in the schedule entity**, not at the point level.

Once the **schedule job** runs, it updates the Unoccupied Setback value **to the point level** as well.
Before the schedule job runs, if we try to check the Unoccupied Setback value, it still shows the **previous (old)** value.

Related work items: #41259