Merged PR 25018: Value sent to modbus device - master
When a point-based schedule is applied, the values update correctly. When Modbus data comes to the CCU, if there’s any difference between the local and device data, the CCU sends back its local value. So, while sending data to the device, equip was not found because of the incorrect query — hence this is now corrected.

Related work items: #42273