Merged PR 25201: Bug 39863: CCU:Zone Time Remains in America/Chicago Despite Changing to Asia/...
Bug 39863: CCU:Zone Time Remains in America/Chicago Despite Changing to Asia/Kolkata During Site Creation

**Observation 1:
LatestModifiedDate and Time not updated when TimeZone is edited**

In the updateSite method in CreateNewSite, the current site details is read,  the current timezone is taken as curTz and then the timeZone gets updated (eg: timezone before editing- Berlin and new Time zone - Melbourne)

Now the updateSite methos is called, in that the Default timeZone is updated, that is the previous time zone as Berlin.

The time zone modified by the user, is captured and is updated using the below given line,
DateTimeZone.setDefault(DateTimeZone.forTimeZone(TimeZone.getTimeZone(timeZone)));

Hence the UI in the Zones Page is getting updated to updated timezone but the lastmodified time zone is set to Default and that is not getting updated

Fix provided:
Hence the HDateTime method is modified to take the systems current timezone as the systems time zone is properly updated when it is changed.

**Observation 2:
When the new site is created and timezone is changed during creation, the saveSite function called doesn't handle the modified time zone.**

Hence the UI didn't get modified according to the modified timezone.

Fix provided:

Hence fix is provided in CreateNewSite in the saveSite method to take the systems time zone before the save operation is completed.

Related work items: #39863